http-server/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
numpy
CaptchaCracker
Pillow
python-multipart
python-dotenv
